Fixed type missmatch on GenericHandlerList constructor

+++ LockLoginCommon/src/main/java/es/karmadev/locklogin/common/util/generic/GenericHandlerList.java
@@ -59,7 +59,7 @@ public final class GenericHandlerList {
      *
      * @param flags the flags
      */
-    public GenericHandlerList(final int flags) {
+    public GenericHandlerList(final byte flags) {
         this.flags = flags;
     }
